refactor(renderer): move views into their own dir "views"
