+++ src/renderer/components/index.ts
@@ -1,3 +1,3 @@
-export * from './MainView'
-export * from './SettingsView'
+export * from './views/MainView'
+export * from './views/SettingsView'
 export * from './SvgIcons'
