+++ src/renderer/components/views/AboutView/logo_64x64.png
